requirements.txt
streamlit 
load_dotenv
fastapi
uvicorn
langchain
langgraph
langchain_groq
langchain_community
langchain_google_genai
langchain-cli
cv2
streamlit
streamlit-webrtc
streamlit-cookies-manager
sqlite3          
bcrypt
python-dotenv
opencv-python
Pillow
av